pyproject.toml
[project]
name = "soundnest-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django-cors-headers>=4.7.0",
    "django-ninja-jwt[crypto]>=5.3.7",
    "django-ninja>=1.4.1",
    "django-ses>=4.4.0",
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "djoser>=2.3.1",
    "pydantic[email]>=2.11.3",
    "python-dotenv>=1.1.0",
    "rav>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/src/soundnest_app"]
